+++ tests/e2e/e2e_gov_test.go
@@ -15,12 +15,12 @@ import (
 	coredaostypes "github.com/atomone-hub/atomone/x/coredaos/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	sdkgovtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 
 	dynamicfeetypes "github.com/atomone-hub/atomone/x/dynamicfee/types"
 	govtypes "github.com/atomone-hub/atomone/x/gov/types"
 	govtypesv1 "github.com/atomone-hub/atomone/x/gov/types/v1"
-	govtypesv1beta1 "github.com/atomone-hub/atomone/x/gov/types/v1beta1"
 	photontypes "github.com/atomone-hub/atomone/x/photon/types"
 )
 
@@ -48,7 +48,7 @@ func (s *IntegrationTestSuite) testGovSoftwareUpgrade() {
 		depositGovFlags := []string{strconv.Itoa(proposalCounter), s.queryGovMinDeposit(chainAAPIEndpoint).String()}
 
 		voteGovFlags := []string{strconv.Itoa(proposalCounter), "yes=0.8,no=0.1,abstain=0.1"}
-		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "SoftwareUpgrade", submitGovFlags, depositGovFlags, voteGovFlags, "weighted-vote", govtypesv1beta1.StatusPassed)
+		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "SoftwareUpgrade", submitGovFlags, depositGovFlags, voteGovFlags, "weighted-vote", govtypesv1.StatusPassed)
 
 		res := s.queryUpgradePlan(chainAAPIEndpoint)
 		s.Require().Equal("v2", res.Plan.Name)
@@ -96,14 +96,14 @@ func (s *IntegrationTestSuite) testGovCancelSoftwareUpgrade() {
 		submitGovFlags := []string{configFile(proposalSoftwareUpgradeFilename)}
 		depositGovFlags := []string{strconv.Itoa(proposalCounter), s.queryGovMinDeposit(chainAAPIEndpoint).String()}
 		voteGovFlags := []string{strconv.Itoa(proposalCounter), "yes"}
-		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "SoftwareUpgrade", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1beta1.StatusPassed)
+		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "SoftwareUpgrade", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1.StatusPassed)
 
 		proposalCounter++
 		s.writeGovCancelUpgradeProposal(s.chainA)
 		submitGovFlags = []string{configFile(proposalCancelUpgradeFilename)}
 		depositGovFlags = []string{strconv.Itoa(proposalCounter), s.queryGovMinDeposit(chainAAPIEndpoint).String()}
 		voteGovFlags = []string{strconv.Itoa(proposalCounter), "yes"}
-		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "CancelUpgrade", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1beta1.StatusPassed)
+		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "CancelUpgrade", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1.StatusPassed)
 
 		s.verifyChainPassesUpgradeHeight(s.chainA, 0, proposalHeight)
 		s.T().Logf("Successfully canceled upgrade at height %d", proposalHeight)
@@ -136,7 +136,7 @@ func (s *IntegrationTestSuite) testGovCommunityPoolSpend() {
 		submitGovFlags := []string{configFile(proposalCommunitySpendFilename)}
 		depositGovFlags := []string{strconv.Itoa(proposalCounter), s.queryGovMinDeposit(chainAAPIEndpoint).String()}
 		voteGovFlags := []string{strconv.Itoa(proposalCounter), "yes"}
-		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "CommunityPoolSpend", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1beta1.StatusPassed)
+		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "CommunityPoolSpend", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1.StatusPassed)
 
 		// Check that sender is refunded with the proposal deposit
 		s.Require().Eventually(
@@ -175,9 +175,9 @@ func (s *IntegrationTestSuite) testGovCommunityPoolSpend() {
 		// Gov tests may be run in arbitrary order, each test must increment proposalCounter to have the correct proposal id to submit and query
 		proposalCounter++
 		submitGovFlags := []string{configFile(proposalCommunitySpendFilename)}
-		depositGovFlags := []string{strconv.Itoa(proposalCounter), s.queryGovMinDeposit(chainAAPIEndpoint).String()}
+		depositGovFlags := []string{strconv.Itoa(proposalCounter), deposit.String()}
 		voteGovFlags := []string{strconv.Itoa(proposalCounter), "no"}
-		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "CommunityPoolSpend", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1beta1.StatusRejected)
+		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "CommunityPoolSpend", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1.StatusRejected)
 
 		// Check that sender is not refunded with the proposal deposit
 		s.Require().Eventually(
@@ -220,7 +220,7 @@ func (s *IntegrationTestSuite) testGovParamChange() {
 		submitGovFlags := []string{configFile(proposalParamChangeFilename)}
 		depositGovFlags := []string{strconv.Itoa(proposalCounter), s.queryGovMinDeposit(chainAAPIEndpoint).String()}
 		voteGovFlags := []string{strconv.Itoa(proposalCounter), "yes"}
-		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "cosmos.staking.v1beta1.MsgUpdateParams", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1beta1.StatusPassed)
+		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "cosmos.staking.v1beta1.MsgUpdateParams", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1.StatusPassed)
 
 		newParams := s.queryStakingParams(chainAAPIEndpoint)
 		s.Assert().NotEqual(oldMaxValidator, newParams.Params.MaxValidators)
@@ -243,7 +243,7 @@ func (s *IntegrationTestSuite) testGovParamChange() {
 		submitGovFlags := []string{configFile(proposalParamChangeFilename)}
 		depositGovFlags := []string{strconv.Itoa(proposalCounter), s.queryGovMinDeposit(chainAAPIEndpoint).String()}
 		voteGovFlags := []string{strconv.Itoa(proposalCounter), "yes"}
-		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "atomone.photon.v1.MsgUpdateParams", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1beta1.StatusPassed)
+		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "atomone.photon.v1.MsgUpdateParams", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1.StatusPassed)
 
 		newParams := s.queryPhotonParams(chainAAPIEndpoint)
 		s.Assert().True(newParams.Params.MintDisabled, "expected photon param mint disabled to be true")
@@ -254,7 +254,7 @@ func (s *IntegrationTestSuite) testGovParamChange() {
 		depositGovFlags = []string{strconv.Itoa(proposalCounter), s.queryGovMinDeposit(chainAAPIEndpoint).String()}
 		voteGovFlags = []string{strconv.Itoa(proposalCounter), "yes"}
 		s.writePhotonParamChangeProposal(s.chainA, params.Params)
-		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "atomone.photon.v1.MsgUpdateParams", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1beta1.StatusPassed)
+		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "atomone.photon.v1.MsgUpdateParams", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1.StatusPassed)
 
 		newParams = s.queryPhotonParams(chainAAPIEndpoint)
 		s.Require().False(newParams.Params.MintDisabled, "expected photon param mint disabled to be false")
@@ -275,7 +275,7 @@ func (s *IntegrationTestSuite) testGovParamChange() {
 		submitGovFlags := []string{configFile(proposalParamChangeFilename)}
 		depositGovFlags := []string{strconv.Itoa(proposalCounter), s.queryGovMinDeposit(chainAAPIEndpoint).String()}
 		voteGovFlags := []string{strconv.Itoa(proposalCounter), "yes"}
-		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "atomone.dynamicfee.v1.MsgUpdateParams", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1beta1.StatusPassed)
+		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "atomone.dynamicfee.v1.MsgUpdateParams", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1.StatusPassed)
 
 		newParams := s.queryDynamicfeeParams(chainAAPIEndpoint)
 		s.Require().Equal(newParams.Params.Alpha, oldAlpha.Add(math.LegacyNewDec(1)))
@@ -297,7 +297,7 @@ func (s *IntegrationTestSuite) testGovConstitutionAmendment() {
 		submitGovFlags := []string{configFile(proposalConstitutionAmendmentFilename)}
 		depositGovFlags := []string{strconv.Itoa(proposalCounter), s.queryGovMinDeposit(chainAAPIEndpoint).String()}
 		voteGovFlags := []string{strconv.Itoa(proposalCounter), "yes"}
-		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "gov/MsgSubmitProposal", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1beta1.StatusPassed)
+		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "gov/MsgSubmitProposal", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1.StatusPassed)
 
 		s.Require().Eventually(
 			func() bool {
@@ -320,7 +320,7 @@ func (s *IntegrationTestSuite) testGovTextProposal() {
 		voteGovFlags := []string{strconv.Itoa(proposalCounter), "yes"}
 		senderAddress, _ := s.chainA.validators[0].keyInfo.GetAddress()
 		sender := senderAddress.String()
-		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "Text", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1beta1.StatusPassed)
+		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "Text", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1.StatusPassed)
 	})
 }
 
@@ -344,14 +344,16 @@ func (s *IntegrationTestSuite) testGovDynamicQuorum() {
 		voteGovFlags := []string{strconv.Itoa(proposalCounter), "yes"}
 		senderAddress, _ := s.chainA.validators[0].keyInfo.GetAddress()
 		sender := senderAddress.String()
-		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "Text", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1beta1.StatusPassed)
+		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "Text", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1.StatusPassed)
 		quorumsAfter := s.queryGovQuorums(chainAAPIEndpoint)
 		endQuorum := math.LegacyMustNewDecFromStr(quorumsAfter.GetQuorum())
 		endQuorumPEma := (endQuorum.Sub(quorumMin)).Quo(quorumMax.Sub(quorumMin))
 		expectedParticipation := endQuorumPEma.Sub(quorumPEma.Mul(math.LegacyMustNewDecFromStr("0.8"))).Quo(math.LegacyMustNewDecFromStr("0.2"))
 		proposal, _ := s.queryGovProposal(chainAAPIEndpoint, proposalCounter)
 		stakingPool := s.queryStakingPool(chainAAPIEndpoint)
-		votes := proposal.Proposal.FinalTallyResult.Yes.ToLegacyDec()
+		yesCount, ok := math.NewIntFromString(proposal.Proposal.FinalTallyResult.YesCount)
+		s.Require().True(ok, "Failed to parse yes count from proposal tally result")
+		votes := yesCount.ToLegacyDec()
 		totalVP := stakingPool.Pool.BondedTokens.ToLegacyDec()
 		actualParticipation := votes.Quo(totalVP)
 
@@ -382,14 +384,16 @@ func (s *IntegrationTestSuite) testGovDynamicQuorum() {
 		sender := senderAddress.String()
 		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter,
 			"gov/MsgSubmitProposal", submitGovFlags, depositGovFlags, voteGovFlags,
-			"vote", govtypesv1beta1.StatusPassed)
+			"vote", govtypesv1.StatusPassed)
 		quorumsAfter := s.queryGovQuorums(chainAAPIEndpoint)
 		endLawQuorum := math.LegacyMustNewDecFromStr(quorumsAfter.GetLawQuorum())
 		endLawQuorumPEma := (endLawQuorum.Sub(lawQuorumMin)).Quo(lawQuorumMax.Sub(lawQuorumMin))
 		expectedParticipation := endLawQuorumPEma.Sub(lawQuorumPEma.Mul(math.LegacyMustNewDecFromStr("0.8"))).Quo(math.LegacyMustNewDecFromStr("0.2"))
 		proposal, _ := s.queryGovProposal(chainAAPIEndpoint, proposalCounter)
 		stakingPool := s.queryStakingPool(chainAAPIEndpoint)
-		votes := proposal.Proposal.FinalTallyResult.Yes.ToLegacyDec()
+		yesCount, ok := math.NewIntFromString(proposal.Proposal.FinalTallyResult.YesCount)
+		s.Require().True(ok, "Failed to parse yes count from proposal tally result")
+		votes := yesCount.ToLegacyDec()
 		totalVP := stakingPool.Pool.BondedTokens.ToLegacyDec()
 		actualParticipation := votes.Quo(totalVP)
 
@@ -420,14 +424,16 @@ func (s *IntegrationTestSuite) testGovDynamicQuorum() {
 		voteGovFlags := []string{strconv.Itoa(proposalCounter), "yes"}
 		senderAddress, _ := s.chainA.validators[0].keyInfo.GetAddress()
 		sender := senderAddress.String()
-		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "gov/MsgSubmitProposal", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1beta1.StatusPassed)
+		s.submitGovProposal(chainAAPIEndpoint, sender, proposalCounter, "gov/MsgSubmitProposal", submitGovFlags, depositGovFlags, voteGovFlags, "vote", govtypesv1.StatusPassed)
 		quorumsAfter := s.queryGovQuorums(chainAAPIEndpoint)
 		endConstitutionAmendmentQuorum := math.LegacyMustNewDecFromStr(quorumsAfter.GetConstitutionAmendmentQuorum())
 		endConstitutionAmendmentQuorumPEma := (endConstitutionAmendmentQuorum.Sub(constitutionAmendmentQuorumMin)).Quo(constitutionAmendmentQuorumMax.Sub(constitutionAmendmentQuorumMin))
 		expectedParticipation := endConstitutionAmendmentQuorumPEma.Sub(constitutionAmendmentQuorumPEma.Mul(math.LegacyMustNewDecFromStr("0.8"))).Quo(math.LegacyMustNewDecFromStr("0.2"))
 		proposal, _ := s.queryGovProposal(chainAAPIEndpoint, proposalCounter)
 		stakingPool := s.queryStakingPool(chainAAPIEndpoint)
-		votes := proposal.Proposal.FinalTallyResult.Yes.ToLegacyDec()
+		yesCount, ok := math.NewIntFromString(proposal.Proposal.FinalTallyResult.YesCount)
+		s.Require().True(ok, "Failed to parse yes count from proposal tally result")
+		votes := yesCount.ToLegacyDec()
 		totalVP := stakingPool.Pool.BondedTokens.ToLegacyDec()
 		actualParticipation := votes.Quo(totalVP)
 
@@ -437,16 +443,120 @@ func (s *IntegrationTestSuite) testGovDynamicQuorum() {
 	})
 }
 
+// testGovGovernors tests passing a text proposal and vote with governors.
+func (s *IntegrationTestSuite) testGovGovernors() {
+	s.Run("governors", func() {
+		chainAAPIEndpoint := fmt.Sprintf("http://%s", s.valResources[s.chainA.id][0].GetHostPort("1317/tcp"))
+		senderAddress, _ := s.chainA.validators[0].keyInfo.GetAddress()
+		sender := senderAddress.String()
+		s.writeGovTextProposal(s.chainA)
+
+		// create a governor
+		acc1Addr, _ := s.chainA.genesisAccounts[1].keyInfo.GetAddress()
+		governorAddr := sdkgovtypes.GovernorAddress(acc1Addr).String()
+		// a governor must have a delegation of at least 10atone
+		valAddr, _ := s.chainA.validators[0].keyInfo.GetAddress()
+		validatorAddr := sdk.ValAddress(valAddr).String()
+		govDelegatorAddr := acc1Addr.String()
+		govDelegation := sdk.NewInt64Coin("uatone", 10_000_000)
+		s.execDelegate(s.chainA, 0, govDelegation, validatorAddr, govDelegatorAddr)
+		// run create-governor
+		s.runGovExec(s.chainA, 0, govDelegatorAddr, "create-governor", []string{
+			govDelegatorAddr, "moniker", "identity", "website", "security-contact", "details",
+		})
+		// check governor is created
+		s.Require().Eventually(
+			func() bool {
+				governor, err := s.queryGovGovernor(chainAAPIEndpoint, governorAddr)
+				s.Require().NoError(err)
+				return governor.Governor != nil
+			},
+			15*time.Second,
+			time.Second,
+		)
+
+		// delegate to this governor
+		// first create a delegator
+		acc2Addr, _ := s.chainA.genesisAccounts[2].keyInfo.GetAddress()
+		delegatorAddr := acc2Addr.String()
+		delDelegation := sdk.NewInt64Coin("uatone", 10_000_000)
+		s.execDelegate(s.chainA, 0, delDelegation, validatorAddr, delegatorAddr)
+		// then delegate to governor
+		s.runGovExec(s.chainA, 0, delegatorAddr, "delegate-governor", []string{
+			delegatorAddr, governorAddr,
+		})
+		// check governor delegation is created
+		s.Require().Eventually(
+			func() bool {
+				resp, err := s.queryGovGovernorDelegation(chainAAPIEndpoint, delegatorAddr)
+				s.Require().NoError(err)
+				return resp.GovernorAddress == governorAddr
+			},
+			15*time.Second,
+			time.Second,
+		)
+		// assert governor valshares
+		resp, err := s.queryGovGovernorValShares(chainAAPIEndpoint, governorAddr)
+		s.Require().NoError(err)
+		s.Require().Len(resp.ValShares, 1, "expected 1 valshare")
+		s.Require().Equal(governorAddr, resp.ValShares[0].GovernorAddress)
+		s.Require().Equal(validatorAddr, resp.ValShares[0].ValidatorAddress)
+		validator, err := s.queryValidator(chainAAPIEndpoint, validatorAddr)
+		s.Require().NoError(err)
+		totalDelegations := delDelegation.Add(govDelegation)
+		expectedShares, err := validator.SharesFromTokens(totalDelegations.Amount)
+		s.Require().NoError(err)
+		s.Require().True(expectedShares.Equal(resp.ValShares[0].Shares), "want shares %s, got %s", expectedShares, resp.ValShares[0].Shares)
+
+		// Create a governance proposal
+		proposalCounter++
+		submitGovFlags := []string{configFile(proposalTextFilename)}
+		s.submitGovCommand(chainAAPIEndpoint, sender, proposalCounter, "submit-proposal", submitGovFlags, govtypesv1.StatusVotingPeriod)
+
+		// Vote with governor
+		voteFlags := []string{strconv.Itoa(proposalCounter), "yes"}
+		s.submitGovCommand(chainAAPIEndpoint, govDelegatorAddr, proposalCounter, "vote", voteFlags, govtypesv1.StatusRejected)
+
+		// assert tally result
+		prop, err := s.queryGovProposal(chainAAPIEndpoint, proposalCounter)
+		s.Require().NoError(err)
+		expectedTally := &govtypesv1.TallyResult{
+			YesCount:     totalDelegations.Amount.String(),
+			NoCount:      "0",
+			AbstainCount: "0",
+		}
+		s.Require().Equal(expectedTally, prop.Proposal.FinalTallyResult)
+	})
+}
+
+func (s *IntegrationTestSuite) submitLegacyGovProposal(chainAAPIEndpoint, sender string, proposalID int, proposalType string, submitFlags []string, depositFlags []string, voteFlags []string, voteCommand string, withDeposit bool) {
+	s.T().Logf("Submitting Gov Proposal: %s", proposalType)
+	// min deposit of 1000uatone is required in e2e tests, otherwise the gov antehandler causes the proposal to be dropped
+	sflags := submitFlags
+	initialDeposit := s.queryGovMinInitialDeposit(chainAAPIEndpoint)
+	if withDeposit {
+		sflags = append(sflags, "--deposit="+initialDeposit.String())
+	}
+	deposit := s.queryGovMinDeposit(chainAAPIEndpoint)
+	depositString := deposit.String()
+	depositFlags = append(depositFlags, depositString)
+	s.submitGovCommand(chainAAPIEndpoint, sender, proposalID, "submit-legacy-proposal", sflags, govtypesv1.StatusDepositPeriod)
+	s.T().Logf("Depositing Gov Proposal: %s", proposalType)
+	s.submitGovCommand(chainAAPIEndpoint, sender, proposalID, "deposit", depositFlags, govtypesv1.StatusVotingPeriod)
+	s.T().Logf("Voting Gov Proposal: %s", proposalType)
+	s.submitGovCommand(chainAAPIEndpoint, sender, proposalID, voteCommand, voteFlags, govtypesv1.StatusPassed)
+}
+
 // NOTE: in SDK >= v0.47 the submit-proposal does not have a --deposit flag
 // Instead, the deposit is added to the "deposit" field of the proposal JSON (usually stored as a file)
 // you can use `atomoned tx gov draft-proposal` to create a proposal file that you can use
 // min initial deposit of 100uatone is required in e2e tests, otherwise the proposal would be dropped
-func (s *IntegrationTestSuite) submitGovProposal(chainAAPIEndpoint, sender string, proposalID int, proposalType string, submitFlags []string, depositFlags []string, voteFlags []string, voteCommand string, expectedStatusAfterVote govtypesv1beta1.ProposalStatus) {
+func (s *IntegrationTestSuite) submitGovProposal(chainAAPIEndpoint, sender string, proposalID int, proposalType string, submitFlags []string, depositFlags []string, voteFlags []string, voteCommand string, expectedStatusAfterVote govtypesv1.ProposalStatus) {
 	s.T().Logf("Submitting Gov Proposal: %s", proposalType)
 	sflags := submitFlags
-	s.submitGovCommand(chainAAPIEndpoint, sender, proposalID, "submit-proposal", sflags, govtypesv1beta1.StatusDepositPeriod)
+	s.submitGovCommand(chainAAPIEndpoint, sender, proposalID, "submit-proposal", sflags, govtypesv1.StatusDepositPeriod)
 	s.T().Logf("Depositing Gov Proposal: %s", proposalType)
-	s.submitGovCommand(chainAAPIEndpoint, sender, proposalID, "deposit", depositFlags, govtypesv1beta1.StatusVotingPeriod)
+	s.submitGovCommand(chainAAPIEndpoint, sender, proposalID, "deposit", depositFlags, govtypesv1.StatusVotingPeriod)
 	s.T().Logf("Voting Gov Proposal: %s", proposalType)
 	s.submitGovCommand(chainAAPIEndpoint, sender, proposalID, voteCommand, voteFlags, expectedStatusAfterVote)
 }
@@ -493,7 +603,7 @@ func (s *IntegrationTestSuite) verifyChainPassesUpgradeHeight(c *chain, valIdx i
 	)
 }
 
-func (s *IntegrationTestSuite) submitGovCommand(chainAAPIEndpoint, sender string, proposalID int, govCommand string, proposalFlags []string, expectedStatus govtypesv1beta1.ProposalStatus) {
+func (s *IntegrationTestSuite) submitGovCommand(chainAAPIEndpoint, sender string, proposalID int, govCommand string, proposalFlags []string, expectedStatus govtypesv1.ProposalStatus) {
 	s.runGovExec(s.chainA, 0, sender, govCommand, proposalFlags)
 
 	s.T().Logf("Waiting for proposal status %s", expectedStatus.String())
@@ -624,6 +734,23 @@ func (s *IntegrationTestSuite) writePhotonParamChangeProposal(c *chain, params p
 	initialDeposit := s.queryGovMinInitialDeposit(chainAAPIEndpoint)
 	propMsgBody := fmt.Sprintf(template, govModuleAddress, s.cdc.MustMarshalJSON(&params), initialDeposit)
 	err := writeFile(filepath.Join(c.validators[0].configDir(), "config", proposalParamChangeFilename), []byte(propMsgBody))
+	s.Require().NoError(err)
+}
+
+// writeGovTextProposal creates a text proposal JSON file with full required deposit.
+func (s *IntegrationTestSuite) writeGovTextProposal(c *chain) {
+	template := `
+	{
+		"deposit": "%s",
+		"metadata": "The metadata",
+		"title": "A text proposal",
+		"summary": "The summary"
+	}
+	`
+	chainAAPIEndpoint := fmt.Sprintf("http://%s", s.valResources[c.id][0].GetHostPort("1317/tcp"))
+	deposit := s.queryGovMinDeposit(chainAAPIEndpoint)
+	propMsgBody := fmt.Sprintf(template, deposit)
+	err := writeFile(filepath.Join(c.validators[0].configDir(), "config", proposalTextFilename), []byte(propMsgBody))
 	s.Require().NoError(err)
 }