+++ tests/e2e/e2e_test.go
@@ -65,6 +65,7 @@ func (s *IntegrationTestSuite) TestGov() {
 	s.testGovConstitutionAmendment()
 	s.testGovDynamicQuorum()
 	s.testGovTextProposal()
+	s.testGovGovernors()
 }
 
 func (s *IntegrationTestSuite) TestIBC_hermesRelayer() {