+++ tests/e2e/e2e_vesting_test.go
@@ -61,8 +61,8 @@ func (s *IntegrationTestSuite) testDelayedVestingAccount(api string) {
 		s.Require().Equal(vestingBalance.AmountOf(uatoneDenom), balance.Amount)
 
 		// Delegate coins should succeed
-		s.execDelegate(chain, valIdx, vestingDelegationAmount.String(), valOpAddr,
-			vestingDelayedAcc.String(), atomoneHomePath)
+		s.execDelegate(chain, valIdx, vestingDelegationAmount, valOpAddr,
+			vestingDelayedAcc.String())
 
 		// Validate delegation successful
 		s.Require().Eventually(
@@ -125,8 +125,8 @@ func (s *IntegrationTestSuite) testContinuousVestingAccount(api string) {
 		s.Require().Equal(vestingBalance.AmountOf(uatoneDenom), balance.Amount)
 
 		// Delegate coins should succeed
-		s.execDelegate(chain, valIdx, vestingDelegationAmount.String(),
-			valOpAddr, continuousVestingAcc.String(), atomoneHomePath)
+		s.execDelegate(chain, valIdx, vestingDelegationAmount,
+			valOpAddr, continuousVestingAcc.String())
 
 		// Validate delegation successful
 		s.Require().Eventually(
@@ -253,8 +253,8 @@ func (s *IntegrationTestSuite) testPeriodicVestingAccount(api string) { //nolint
 		}
 
 		// Delegate coins should succeed
-		s.execDelegate(chain, valIdx, vestingDelegationAmount.String(), valOpAddr,
-			periodicVestingAddr, atomoneHomePath)
+		s.execDelegate(chain, valIdx, vestingDelegationAmount, valOpAddr,
+			periodicVestingAddr)
 
 		// Validate delegation successful
 		s.Require().Eventually(